Avoid signed unsigned comparisons
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@5426 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/thd.cpp
@@ -1365,7 +1365,7 @@ int THD_rec_mutex_lock(REC_MUTX_T * rec_mutex)
  **************************************/
 	int ret;
 
-	if (rec_mutex->rec_mutx_id == THD_get_thread_id())
+	if (rec_mutex->rec_mutx_id == (ULONG) THD_get_thread_id())
 		rec_mutex->rec_mutx_count++;
 	else {
 		if (ret = THD_mutex_lock(rec_mutex->rec_mutx_mtx))
@@ -1389,7 +1389,7 @@ int THD_rec_mutex_unlock(REC_MUTX_T * rec_mutex)
  *
  **************************************/
 
-	if (rec_mutex->rec_mutx_id != THD_get_thread_id())
+	if (rec_mutex->rec_mutx_id != (ULONG) THD_get_thread_id())
 		return FB_FAILURE;
 
 	rec_mutex->rec_mutx_count--;
